+++ GRDB Tests/Public/Core/DatabasePersistable/MutableDatabasePersistableTests.swift
@@ -63,9 +63,9 @@ struct MutablePersistableCustomizedCountry : MutableDatabasePersistable {
         try performInsert(db)
     }
     
-    func update(db: Database) throws {
+    func update(db: Database, columns: [String]? = nil) throws {
         willUpdate()
-        try performUpdate(db)
+        try performUpdate(db, columns: columns)
     }
     
     mutating func save(db: Database) throws {