+++ GRDB Tests/Public/Record/RecordSubClassTests.swift
@@ -37,9 +37,9 @@ class PersonWithOverrides : Person {
         try super.insert(db)
     }
     
-    override func update(db: Database) throws {
+    override func update(db: Database, columns: [String]? = nil) throws {
         lastSavingMethod = .Update
-        try super.update(db)
+        try super.update(db, columns: columns)
     }
 }